+++ UD22_Ejercicio_1/ud22_ejercicio1/src/main/java/com/example/ud22_models/ClienteClass.java
@@ -2,23 +2,121 @@ package com.example.ud22_models;
 
 import java.util.Date;
 
+/**
+ * @author Josep Oriol López Bosch
+ * @author David Dalmau Dieguez
+ * @author Octavio Bernal Vilana
+ */
 public class ClienteClass {
-	
+
+	/**
+	 * Se crean los atributos de clase correspondiente a las columnas de la tablas.
+	 */
 	private int id;
 	private String nombre;
 	private String apellido;
 	private String direccion;
-	private int dni;
+	private String dni;
 	private Date fecha;
-	
-	//Constructor without id (id should be controled by the dataBase Server)
-	public ClienteClass(String nombre, String apellido, String direccion, int dni, Date fecha) {
-		super();
+
+	/**
+	 * Se crea constructor vacío
+	 */
+	public ClienteClass() {
+	}
+
+	/**
+	 * Se crea constructor completo
+	 * 
+	 * @param nombre
+	 * @param apellido
+	 * @param direccion
+	 * @param fecha
+	 */
+	public ClienteClass(String nombre, String apellido, String direccion, String dni, Date fecha) {
 		this.nombre = nombre;
 		this.apellido = apellido;
 		this.direccion = direccion;
 		this.dni = dni;
 		this.fecha = fecha;
 	}
-	
+
+	// Se crean los getters y setters de las diferentes columnas.
+	/**
+	 * @return the id
+	 */
+	public int getId() {
+		return id;
+	}
+
+	/**
+	 * @return the nombre
+	 */
+	public String getNombre() {
+		return nombre;
+	}
+
+	/**
+	 * @param nombre the nombre to set
+	 */
+	public void setNombre(String nombre) {
+		this.nombre = nombre;
+	}
+
+	/**
+	 * @return the apellido
+	 */
+	public String getApellido() {
+		return apellido;
+	}
+
+	/**
+	 * @param apellido the apellido to set
+	 */
+	public void setApellido(String apellido) {
+		this.apellido = apellido;
+	}
+
+	/**
+	 * @return the direccion
+	 */
+	public String getDireccion() {
+		return direccion;
+	}
+
+	/**
+	 * @param direccion the direccion to set
+	 */
+	public void setDireccion(String direccion) {
+		this.direccion = direccion;
+	}
+
+	/**
+	 * @return the dni
+	 */
+	public String getDni() {
+		return dni;
+	}
+
+	/**
+	 * @param dni to set to dni
+	 */
+	public void setDni() {
+		this.dni = dni;
+	}
+
+	/**
+	 * @return the fecha
+	 */
+	public Date getFecha() {
+		return fecha;
+	}
+
+	/**
+	 * @param fecha the fecha to set
+	 */
+	public void setFecha(Date fecha) {
+		this.fecha = fecha;
+	}
+
 }
