+++ UD22_Ejercicio_1/ud22_ejercicio1/src/main/java/com/example/ud22_models/ModelFunctions.java
@@ -20,7 +20,7 @@ public class ModelFunctions {
 	static String userSQL;
 	static String password;
 
-	ModelDDBB modelo = new ModelDDBB();
+	ClienteClass modelo = new ClienteClass();
 
 	/**
 	 * Función para seleccionar el servidor.
